server/requirements.txt
fastapi==0.115.8
uvicorn==0.29.0
openai==1.65.1
httpx==0.27.0
bs4==0.0.2
pocketbase-async==0.12.0
pydantic==2.11.7
